Fix bugs

+++ API/Rust/src/definitions.rs
@@ -1,22 +1,22 @@
 pub struct Transport {
     pub name: String,
-    pub geometry: f32,
-    pub ljpotential: f32,
-    pub ljdiameter: f32,
-    pub dipole: f32,
-    pub polar: f32,
-    pub collision: f32,
-    pub mw: f32
+    pub geometry: f64,
+    pub ljpotential: f64,
+    pub ljdiameter: f64,
+    pub dipole: f64,
+    pub polar: f64,
+    pub collision: f64,
+    pub mw: f64
 }
 
 pub struct Thermo {
     pub name: String,
-    pub high: [f32; 7],
-    pub low: [f32; 7],
+    pub high: [f64; 7],
+    pub low: [f64; 7],
 }
 
 pub struct Omega {
-    pub d: [f32; 8],
-    pub t: [f32; 37],
-    pub sigma: [[f32; 8]; 37],
+    pub d: [f64; 8],
+    pub t: [f64; 37],
+    pub sigma: [[f64; 8]; 37],
 }
